#3 laser damage

+++ Assets/Scripts/Enemy.cs
@@ -6,7 +6,7 @@ public class Enemy : MonoBehaviour
 {
     public float speed = 10f;
 
-    public int health = 100;
+    public float health = 100;
 
     public int value = 50;
 
@@ -21,7 +21,7 @@ public class Enemy : MonoBehaviour
         target = Waypoints.points[0];
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         health -= amount;
